refactor: separate article components
