+++ src/components/Article/Article.module.css
@@ -1,7 +1,3 @@
-.relatedArticles > div > button {
-  border-top: none !important;
-}
-
 /* temporary breakpoint */
 @media (min-width: 1024px) {
   .mobileToc {
