+++ src/components/Article/index.ts
@@ -1,1 +1,1 @@
-export { default as Article } from './Article'
+export { default as Article } from './Article.Body'
